// api/src/utils/db/schemas/COMPANY.ts
export interface Company {
	id: number;
	company_name: string;
	company_address: string;
	company_postal_code: string;
	company_email?: string;
	company_phone?: string;
	company_website?: string;
	company_tax_number: string;
	company_vat_id: string;
	company_director: string;
	commercial_register: string;
	bank_name: string;
	company_iban: string;
	company_bic: string;
}